+++ packages/learn-card-base/src/helpers/walletHelpers.ts
@@ -12,10 +12,10 @@ import type { BespokeLearnCard } from 'learn-card-base/types/learn-card';
 import { switchedProfileStore, walletStore } from 'learn-card-base/stores/walletStore';
 import { isPlatformWeb } from 'learn-card-base/helpers/platformHelpers';
 import { requireCurrentUserPrivateKey } from 'learn-card-base/helpers/privateKeyHelpers';
-import { LEARNCARD_NETWORK_URL } from 'learn-card-base/constants/Networks';
 import { networkStore } from 'learn-card-base/stores/NetworkStore';
 import { QueryClient } from '@tanstack/react-query';
 import { getGuardianApprovalVP } from 'learn-card-base/stores/guardianApprovalStore';
+import { PRODUCTION_NETWORK_URL } from './networkHelpers';
 
 let LEARN_CARDS: Record<string, BespokeLearnCard> = {};
 
@@ -67,7 +67,7 @@ export const getBespokeLearnCard = async (
     if (LEARN_CARDS[cacheKey]) return LEARN_CARDS[cacheKey];
 
     let network: string | boolean = networkStore.get.networkUrl();
-    if (!network || network === LEARNCARD_NETWORK_URL) network = true;
+    if (!network || network === PRODUCTION_NETWORK_URL) network = true;
 
     const cloudUrl = networkStore.get.cloudUrl();